moved code for user modifications from main to login blueprint
